Arreglo error pytest
